java Flux Control

+++ java/initial/src/initial/calculadora/Calculadora.java
@@ -18,7 +18,7 @@ public class Calculadora {
 		int soma = soma(a,b);
 		int subtracao = subtracao(a,b);
 		int multiplicacao = multiplicacao(a,b);
-		double divisao = divisao(a,b);
+		float divisao = divisao(a,b);
 		
 		System.out.println("soma " + soma);
 		System.out.println("subtracao " + subtracao);
@@ -41,7 +41,7 @@ public class Calculadora {
 		
 	}
 	
-	public static double divisao(int a, int b) {
+	public static float divisao(int a, int b) {
 		return a / b;
 		
 	}
